requirements.txt
ahk==1.1.3
customtkinter==5.1.3
Pillow==9.5.0
PyAutoGUI==0.9.54
pytube==15.0.0
Requests==2.31.0
pytubefix